fix: Add math.h include and fix uint8_t overflow
- Add #include <math.h> for sinf() in emergency.c
- Change activation_duration_ms from uint8_t to uint16_t (1000ms > 255)

+++ main/comfort/emergency.h
@@ -26,7 +26,7 @@ typedef enum {
  */
 typedef struct {
     bool enabled;                     // Enable emergency calm feature
-    uint8_t activation_duration_ms;   // How long to hold both buttons (default: 1000)
+    uint16_t activation_duration_ms;  // How long to hold both buttons (default: 1000)
     uint8_t auto_exit_minutes;        // Auto-exit after X minutes (default: 2)
     bool haptic_enabled;              // Enable haptic feedback
     bool audio_enabled;               // Enable soothing audio (future)
